+++ libs/search-widget/src/core/search-bar.ts
@@ -17,7 +17,9 @@ export const setupTriggerSearch = (dispatch: (event: string, details: any) => vo
       switchMap((query) =>
         forkJoin([nucliaStore().searchOptions.pipe(take(1)), nucliaStore().filters.pipe(take(1))]).pipe(
           map(([options, filters]) => {
-            const show = navigateToLink.getValue() ? [ResourceProperties.BASIC, ResourceProperties.VALUES] : [];
+            const show = navigateToLink.getValue()
+              ? [ResourceProperties.BASIC, ResourceProperties.VALUES]
+              : [ResourceProperties.BASIC];
             const currentOptions = { ...options, show, filters };
             return { query, options: currentOptions };
           }),