+++ libs/search-widget/src/old-components/row/Row.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-  import { IResource, ReadableResource } from '@nuclia/core';
+  import { ReadableResource, Search } from '@nuclia/core';
   import { formatDate, formatTitle } from '../../core/utils';
   import { _ } from '../../core/i18n';
-  import { nucliaState } from '../../core/old-stores/main.store';
   import MimeIcon from '../../common/icons/MimeIcon.svelte';
   import Thumbnail from '../../common/thumbnail/Thumbnail.svelte';
+  import AllResultsToggle from '../../common/paragraph-result/AllResultsToggle.svelte';
   import Label from '../../common/label/Label.svelte';
   import Dropdown from '../../common/dropdown/Dropdown.svelte';
   import Button from '../../common/button/Button.svelte';
@@ -13,10 +13,10 @@
 
   export let displayThumbnail = true;
   export let formWidget = false;
-  export let result: IResource;
-  export let semantic = false;
-  const paragraphs = nucliaState().getMatchingParagraphs(result.id);
-  const sentences = nucliaState().getMatchingSentences(result.id);
+  export let result: Search.SmartResult;
+  const paragraphs = result.paragraphs || [];
+  let showAllResults = false;
+
   $: labels = new ReadableResource(result).getClassifications();
 
   const labelDisplayLimit = 4;
@@ -53,29 +53,29 @@
     </div>
     <div class="block-2">
       <h2 class="title">{formatTitle(result.title)}</h2>
-      {#if semantic && $sentences.length > 0}
+      {#if paragraphs.length > 0}
         <ul class="paragraph-list">
-          {#each $sentences as sentence}
-            <li
-              class="paragraph"
-              on:click|preventDefault|stopPropagation={() =>
-                goToResource({ uid: sentence.rid, sentence }, sentence.text)}>
-              {@html sentence.text}
-            </li>
+          {#each paragraphs as paragraph, index}
+            {@const isSemantic = paragraph.sentences && paragraph.sentences.length > 0}
+            {#if index < 4 || showAllResults}
+              <li
+                class="paragraph"
+                class:semantic={isSemantic}
+                on:click|preventDefault|stopPropagation={() =>
+                  goToResource(
+                    { uid: paragraph.rid, [isSemantic ? 'sentence' : 'paragraph']: paragraph },
+                    paragraph.text,
+                  )}>
+                {@html paragraph.text}
+              </li>
+            {/if}
           {/each}
         </ul>
-      {/if}
-      {#if !semantic && $paragraphs.length > 0}
-        <ul class="paragraph-list">
-          {#each $paragraphs as paragraph}
-            <li
-              class="paragraph"
-              on:click|preventDefault|stopPropagation={() =>
-                goToResource({ uid: paragraph.rid, paragraph }, paragraph.text)}>
-              {@html paragraph.text}
-            </li>
-          {/each}
-        </ul>
+        {#if paragraphs.length > 4}
+          <AllResultsToggle
+            {showAllResults}
+            on:toggle={() => (showAllResults = !showAllResults)} />
+        {/if}
       {/if}
     </div>
     {#if displayThumbnail}