+++ libs/search-widget/src/tiles/video-tile/VideoTile.svelte
@@ -14,7 +14,7 @@
     getMainFieldParagraphs,
     getVideoStream,
   } from '../../core/old-stores/viewer.store';
-  import { FieldType, MediaWidgetParagraph } from '../../core/models';
+  import { FieldType, MediaWidgetParagraph, PreviewKind } from '../../core/models';
   import Icon from '../../common/icons/Icon.svelte';
   import { fade, slide } from 'svelte/transition';
   import Player from '../../old-components/viewer/previewers/Player.svelte';
@@ -23,8 +23,9 @@
   import AllResultsToggle from '../../common/paragraph-result/AllResultsToggle.svelte';
   import { _ } from '../../core/i18n';
   import { freezeBackground, unblockBackground } from '../../common/modal/modal.utils';
+  import { mapSmartParagraph2WidgetParagraph } from '../../core/utils';
 
-  export let result: IResource = { id: '' } as IResource;
+  export let result: Search.SmartResult = { id: '' } as Search.SmartResult;
 
   let innerWidth = window.innerWidth;
 
@@ -46,22 +47,8 @@
   let showFullTranscripts = false;
   let animatingShowFullTranscript = false;
 
-  const matchingParagraphs = nucliaState()
-    .getMatchingParagraphs(result.id)
-    .pipe(
-      map((results) =>
-        results.map((paragraph) => {
-          const start = (paragraph as Paragraph).start_seconds?.[0] || 0;
-          const end = (paragraph as Paragraph).end_seconds?.[0] || 0;
-          return {
-            ...paragraph,
-            pid: `${paragraph.field_type}${start}${end}`,
-            start_seconds: start,
-            end_seconds: end,
-          };
-        }),
-      ),
-    );
+  const matchingParagraphs: MediaWidgetParagraph[] =
+    result.paragraphs?.map((paragraph) => mapSmartParagraph2WidgetParagraph(paragraph, PreviewKind.VIDEO)) || [];
 
   const filterParagraphs = (paragraphs: MediaWidgetParagraph[]): MediaWidgetParagraph[] => {
     return paragraphs
@@ -77,16 +64,14 @@
 
   $: isMobile = innerWidth < 448;
   $: defaultTransitionDuration = expanded ? Duration.MODERATE : 0;
-  $: filteredMatchingParagraphs = !findInTranscript
-    ? matchingParagraphs
-    : matchingParagraphs.pipe(map((paragraphs) => filterParagraphs(paragraphs as MediaWidgetParagraph[])));
+  $: filteredMatchingParagraphs = !findInTranscript ? matchingParagraphs : filterParagraphs(matchingParagraphs || []);
   $: filteredTranscripts = !findInTranscript ? transcripts : filterParagraphs(transcripts);
 
   const playFromStart = () => {
     playFrom(0);
   };
 
-  const playParagraph = (paragraph) => {
+  const playParagraph = (paragraph: MediaWidgetParagraph) => {
     playFrom(paragraph.start_seconds, paragraph);
     paragraphInPlay = paragraph;
   };
@@ -100,39 +85,31 @@
     return fieldType === FieldType.FILE || fieldType === FieldType.LINK;
   };
 
-  const playFrom = (time: number, selectedParagraph?: Search.Paragraph) => {
+  const playFrom = (time: number, selectedParagraph?: MediaWidgetParagraph) => {
     videoTime = time;
     if (!expanded) {
       expanded = true;
       freezeBackground(true);
     }
 
-    const paragraph$ =
-      selectedParagraph && isFileOrLink(selectedParagraph.field_type)
-        ? of(selectedParagraph)
-        : matchingParagraphs.pipe(
-            map((paragraphList) => paragraphList.filter((p) => isFileOrLink(p.field_type))[0] || paragraphList[0]),
-          );
+    const paragraph =
+      selectedParagraph && isFileOrLink(selectedParagraph.fieldType)
+        ? selectedParagraph
+        : matchingParagraphs.filter((p) => isFileOrLink(p.fieldType))[0] || matchingParagraphs[0];
     if (!resource) {
-      resource = paragraph$.pipe(
-        switchMap((paragraph) =>
-          getResource(result.id).pipe(
-            tap((res) => {
-              if (paragraph.field_type === FieldType.LINK) {
-                const linkField = getLinkField(res, paragraph.field);
-                youtubeUri = linkField?.value?.uri;
-              } else if (paragraph.field_type === FieldType.FILE) {
-                const fileField = getFileField(res, res.id);
-                const file = fileField && (getVideoStream(fileField) || fileField.value?.file);
-                if (file) {
-                  videoContentType = file.content_type;
-                  videoUri = `${getRegionalBackend()}${file.uri}`;
-                }
-              }
-            }),
-          ),
-        ),
+      resource = getResource(result.id).pipe(
         tap((res) => {
+          if (paragraph.fieldType === FieldType.LINK) {
+            const linkField = getLinkField(res, paragraph.fieldId);
+            youtubeUri = linkField?.value?.uri;
+          } else if (paragraph.fieldType === FieldType.FILE) {
+            const fileField = getFileField(res, res.id);
+            const file = fileField && (getVideoStream(fileField) || fileField.value?.file);
+            if (file) {
+              videoContentType = file.content_type;
+              videoUri = `${getRegionalBackend()}${file.uri}`;
+            }
+          }
           const summaries = res.summary ? [res.summary] : res.getExtractedSummaries();
           summary = summaries.filter((s) => !!s)[0];
           transcripts = getMainFieldParagraphs(res) as MediaWidgetParagraph[];
@@ -249,7 +226,7 @@
           tabindex="-1"
           class:on-animation={animatingShowFullTranscript}
           class:transcript-container={expanded}>
-          {#if !showFullTranscripts && findInTranscript && $filteredMatchingParagraphs.length === 0}
+          {#if !showFullTranscripts && findInTranscript && filteredMatchingParagraphs.length === 0}
             <strong>{findInTranscript}</strong>
             not found in your search results…
           {/if}
@@ -269,7 +246,7 @@
               <div
                 tabindex="-1"
                 class="transcript-expander-header-title">
-                <strong>{$matchingParagraphs.length} search results</strong>
+                <strong>{matchingParagraphs.length} search results</strong>
               </div>
               <div
                 tabindex="-1"
@@ -284,9 +261,9 @@
               transition:slide={{ duration: defaultTransitionDuration }}
               class="paragraphs-container"
               class:expanded={showAllResults}
-              class:can-expand={$matchingParagraphs.length > 4}
-              style="--paragraph-count: {$matchingParagraphs.length}">
-              {#each $filteredMatchingParagraphs as paragraph}
+              class:can-expand={matchingParagraphs.length > 4}
+              style="--paragraph-count: {matchingParagraphs.length}">
+              {#each filteredMatchingParagraphs as paragraph}
                 <ParagraphResult
                   {paragraph}
                   ellipsis={!expanded}
@@ -341,7 +318,7 @@
         </div>
       </div>
 
-      {#if !expanded && $matchingParagraphs.length > 4}
+      {#if !expanded && matchingParagraphs.length > 4}
         <AllResultsToggle
           {showAllResults}
           on:toggle={() => (showAllResults = !showAllResults)} />